feat: rich text block
