+++ content/pages/example-rich-text.mdx
@@ -1,0 +1,8 @@
+---
+title: Example Rich Text
+blocks:
+  - richTextContent: >
+      Words Words Words <ColoredText color="Ally Green" text="Green stuff" />
+      more words!
+    _template: richText
+---
