+++ src/components/RenderBlocks.tsx
@@ -8,6 +8,7 @@ import { LookingForContainer } from './blocks/LookingFor'
 import { TitleImageGrid } from './blocks/layout/TitleImageGrid'
 import { Header } from './blocks/layout/Header'
 import { TwoColumn } from './blocks/layout/TwoColumn'
+import { RichText } from './blocks/layout/RichText'
 
 export const Blocks = (props: Omit<Page, 'id' | '_sys' | '_values'>) => {
 	return (
@@ -67,6 +68,12 @@ export const Blocks = (props: Omit<Page, 'id' | '_sys' | '_values'>) => {
 								return (
 									<div key={i + block.__typename} className='w-full'>
 										<TwoColumn data={block} />
+									</div>
+								)
+							case 'PageBlocksRichText':
+								return (
+									<div key={i + block.__typename} className='w-full'>
+										<RichText data={block} />
 									</div>
 								)
 						}
